+++ onnxruntime/core/providers/cpu/cpu_execution_provider.cc
@@ -1106,6 +1106,7 @@ Status RegisterOnnxOperatorKernels(KernelRegistry& kernel_registry) {
 }
 
 // Forward declarations of ml op kernels
+#ifndef DISABLE_ML_OPS
 namespace ml {
 class ONNX_OPERATOR_TYPED_KERNEL_CLASS_NAME(kCpuExecutionProvider, kMLDomain, 1, float, ArrayFeatureExtractor);
 class ONNX_OPERATOR_TYPED_KERNEL_CLASS_NAME(kCpuExecutionProvider, kMLDomain, 1, double, ArrayFeatureExtractor);
@@ -1241,6 +1242,7 @@ Status RegisterOnnxMLOperatorKernels(KernelRegistry& kernel_registry) {
   return Status::OK();
 }
 }  // namespace ml
+#endif
 
 static Status RegisterCPUKernels(KernelRegistry& kernel_registry) {
   ORT_RETURN_IF_ERROR(RegisterOnnxOperatorKernels(kernel_registry));
