+++ tools/ci_build/build.py
@@ -1177,6 +1177,9 @@ def run_onnxruntime_tests(args, source_dir, ctest_path, build_dir, configs):
                 [sys.executable, 'onnxruntime_test_python.py'],
                 cwd=cwd, dll_path=dll_path)
 
+            if not args.disable_ml_ops:
+                run_subprocess([sys.executable, 'onnxruntime_test_python_mlops.py'], cwd=cwd, dll_path=dll_path)
+
             if args.enable_training and args.use_cuda:
                 # run basic frontend tests
                 run_training_python_frontend_tests(cwd=cwd)
@@ -1197,6 +1200,12 @@ def run_onnxruntime_tests(args, source_dir, ctest_path, build_dir, configs):
                 run_subprocess(
                     [sys.executable, 'onnxruntime_test_python_backend.py'],
                     cwd=cwd, dll_path=dll_path)
+
+                if not args.disable_ml_ops:
+                    run_subprocess(
+                        [sys.executable, 'onnxruntime_test_python_backend_mlops.py'],
+                        cwd=cwd, dll_path=dll_path)
+
                 run_subprocess(
                     [sys.executable,
                      os.path.join(source_dir, 'onnxruntime', 'test', 'onnx',
